It works !!
If you want to use the "real" EventStream, change the "Dummy" key to
"False" in app.config
you should also run the Azure Storage Emulator.

Performances are really bad, dunno if it's our fault or the ASE ?
(~1 message per second).

But anyway, it works

+++ EvernestBack/Reader.cs
@@ -11,7 +11,7 @@ namespace EvernestBack
     {
         System.IO.Stream readStream;
         public Reader(string Message, long RequestID, System.IO.Stream readStream, Action<IAgent> Callback)
-            :base(Message, (ulong) RequestID, Callback) // TODO : remove ulong cast when Agent updated
+            :base(Message, RequestID, Callback) // TODO : remove ulong cast when Agent updated
         {
             this.readStream = readStream;
         }
